@@ -1,9 +1,9 @@
 {
   "name": "angularfire-lite",
-  "version": "5.2.0",
+  "version": "6.0.2-rc3",
   "license": "Apache-2.0",
   "description": "Lightweight library to use Firebase API with Angular.",
-  "author": "Hamed Baatour",
+  "author": "Hamed Baatour <hamedbaatour@gmail.com>",
   "repository": {
     "type": "git",
     "url": "https://github.com/hamedbaatour/angularfire-lite"
@@ -13,30 +13,31 @@
   },
   "private": false,
   "peerDependencies": {
-    "@angular/common": "^5.2.0",
-    "@angular/compiler": "^5.2.0",
-    "@angular/core": "^5.2.0",
-    "@angular/http": "^5.2.0",
-    "@angular/platform-browser": "^5.2.0",
-    "@angular/platform-browser-dynamic": "^5.2.0",
-    "core-js": "^2.4.1",
-    "firebase": "^4.12.1",
-    "rxjs": "^5.5.6",
-    "zone.js": "^0.8.19"
+    "@angular/common": "^6.0.0",
+    "@angular/compiler": "^6.0.0",
+    "@angular/core": "^6.0.0",
+    "@angular/http": "^6.0.0",
+    "@angular/platform-browser": "^6.0.0",
+    "@angular/platform-browser-dynamic": "^6.0.0",
+    "core-js": "^2.5.4",
+    "firebase": "^5.0.1",
+    "rxjs": "^6.0.0",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "ng-packagr": "^2.4.2",
-    "@angular/common": "^5.2.0",
-    "@angular/compiler": "^5.2.0",
-    "@angular/core": "^5.2.0",
-    "@angular/http": "^5.2.0",
-    "@angular/compiler-cli": "^5.2.0",
-    "@angular/platform-browser": "^5.2.0",
-    "@angular/platform-browser-dynamic": "^5.2.0",
-    "core-js": "^2.4.1",
-    "firebase": "^4.12.1",
-    "rxjs": "^5.5.6",
-    "zone.js": "^0.8.19",
-    "typescript": "~2.5.3"
+    "@angular/common": "^6.0.0",
+    "@angular/compiler": "^6.0.0",
+    "@angular/compiler-cli": "^6.0.0",
+    "@angular/core": "^6.0.0",
+    "@angular/http": "^6.0.0",
+    "@angular/platform-browser": "^6.0.0",
+    "@angular/platform-browser-dynamic": "^6.0.0",
+    "core-js": "^2.5.4",
+    "firebase": "^5.0.4",
+    "ng-packagr": "^3.0.0",
+    "rxjs": "^6.0.0",
+    "tslint": "~5.9.1",
+    "typescript": "~2.7.2",
+    "zone.js": "^0.8.26"
   }
 }
